Exam mode: Rename the test-exam working-time folder and spec to kebab-case
The folder was `testExam-workingTime` and the spec `testexam-working-time
.component.spec.ts`, while the component and template beside them already used
kebab-case. The client convention is kebab-case for filenames, so all four now
agree: test-exam-working-time/test-exam-working-time.component.{ts,html,spec.ts}.

Renamed through a temporary path so git records the case change on a
case-insensitive filesystem, and updated the three import paths that referenced
the old folder.

Verified: the component and all three consumer specs pass (12 files / 152
tests), the whole exam module passes, and `pnpm run compile` and `compile:tests`
both exit 0 — the latter two would fail on any missed import path.
